requirements.txt
pywinauto>=0.6.8
paddlepaddle>=2.4.0
paddleocr>=2.6.0
pillow>=9.0.0
keyboard>=0.13.5
mouse>=0.7.1
